Initial pass. Example runs.
Full disclosure - this commit was assisted with Cursor AI. However, I have
checked each line and have ended up re-writing most of them. The
involvement of the AI was bootstrapping the functional options pattern
and it ended up writing some of the example.go code and the README.
However, I ended up needing to toss most of the work that it did outside
of the preRun and PostRun wrappers as it was mostly irrelevant and
obtuse for this use case.