Adds delta temporality
When pushing metrics like this with cumulative temporality, we do not
get an accurate picture because it drops additional commands run with
the same invocation and still reports a single command run until the
metric rotates out of the collector. With Delta temporality, we report
an additional count metric each time, and when coupled with a collector
that is running the `deltatocumulative` processor, this allows the count
of invocations to increment as additional calls are made, allowing the
correct number to of invocations to be scraped by an exporter.